Move tests related to inetnum allocation that have LIR locked/unlocked attributes in separate class. (can be copied for inet6num)
