+++ whois-endtoend/src/test/groovy/net/ripe/db/whois/spec/update/AllocationAttributeValidationSpec.groovy
@@ -146,61 +146,6 @@ class AllocationAttributeValidationSpec extends BaseQueryUpdateSpec {
         ]
     }
 
-    ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-    //
-    //  LIR *ADD* attributes (inetnum and inet6num)
-    //
-    ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-    def "modify inetnum, add (all) lir-unlocked attributes by lir"() {
-        given:
-        syncUpdate(getTransient("ALLOC-PA-MANDATORY") + "override: denis, override1")
-        syncUpdate(getTransient("IRT") + "override: denis, override1")
-
-        expect:
-        queryObject("-GBr -T inetnum 192.168.0.0 - 192.169.255.255", "inetnum", "192.168.0.0 - 192.169.255.255")
-        queryObject("-r -T irt irt-test", "irt", "irt-test")
-
-        when:
-        def ack = syncUpdateWithResponse("""
-                inetnum:      192.168.0.0 - 192.169.255.255
-                netname:      TEST-NET-NAME
-                descr:        some description  # added
-                country:      NL
-                country:      DE                # added
-                geoloc:       0.0 0.0           # added
-                language:     NL                # added
-                org:          ORG-LIR1-TEST
-                admin-c:      TP1-TEST
-                admin-c:      TP2-TEST          # added
-                tech-c:       TP1-TEST
-                tech-c:       TP2-TEST          # added
-                remarks:      a new remark      # added
-                notify:       notify@ripe.net   # added
-                status:       ALLOCATED PA
-                mnt-by:       RIPE-NCC-HM-MNT
-                mnt-by:       LIR-MNT
-                mnt-lower:    LIR2-MNT          # added
-                mnt-routes:   OWNER-MNT         # added
-                mnt-domains:  DOMAINS-MNT       # added
-                mnt-irt:      IRT-TEST          # added
-                source:       TEST
-                password: lir
-                password: irt
-                """.stripIndent()
-        )
-
-        then:
-        ack.success
-
-        ack.summary.nrFound == 1
-        ack.summary.assertSuccess(1, 0, 1, 0, 0)
-        ack.summary.assertErrors(0, 0, 0, 0)
-
-        ack.countErrorWarnInfo(0, 0, 0)
-        ack.successes.any { it.operation == "Modify" && it.key == "[inetnum] 192.168.0.0 - 192.169.255.255" }
-    }
-
     def "modify inet6num, add mnt-lower with lir password should be possible"() {
         given:
         syncUpdate(getTransient("V6ALLOC-RIR") + "override: denis, override1")
@@ -294,49 +239,6 @@ class AllocationAttributeValidationSpec extends BaseQueryUpdateSpec {
 
     }
 
-    def "modify inetnum, change lir-locked attributes by lir password should not be possible"() {
-        given:
-        syncUpdate(getTransient("ALLOC-PA-MANDATORY") + "override: denis, override1")
-
-        expect:
-        queryObject("-GBr -T inetnum 192.168.0.0 - 192.169.255.255", "inetnum", "192.168.0.0 - 192.169.255.255")
-
-        when:
-        def ack = syncUpdateWithResponse("""
-                inetnum:      192.168.0.0 - 192.169.255.255
-                netname:      TEST-NET-NAME-CHANGED # changed
-                country:      NL
-                org:          ORG-OTO1-TEST         # changed
-                admin-c:      TP1-TEST
-                tech-c:       TP1-TEST
-                status:       ALLOCATED PI          # changed
-                mnt-by:       RIPE-NCC-HM-MNT
-                mnt-by:       LIR2-MNT              # changed
-                source:       TEST
-                password: lir
-                password: owner3
-                """.stripIndent()
-        )
-
-        then:
-        ack.errors
-
-        ack.summary.nrFound == 1
-        ack.summary.assertSuccess(0, 0, 0, 0, 0)
-        ack.summary.assertErrors(1, 0, 1, 0)
-
-        ack.countErrorWarnInfo(5, 0, 0)
-        ack.errors.any { it.operation == "Modify" && it.key == "[inetnum] 192.168.0.0 - 192.169.255.255" }
-
-        ack.errorMessagesFor("Modify", "[inetnum] 192.168.0.0 - 192.169.255.255") == [
-                "Referenced organisation can only be changed by the RIPE NCC for this resource. Please contact \"ncc@ripe.net\" to change this reference.",
-                "Attribute \"mnt-by:\" can only be changed by the RIPE NCC for this object. Please contact \"ncc@ripe.net\" to change it.",
-                "The \"netname\" attribute can only be changed by the RIPE NCC",
-                "Referenced organisation has wrong \"org-type\". Allowed values are [IANA, RIR, LIR]",
-                "status value cannot be changed, you must delete and re-create the object"
-        ]
-    }
-
     def "modify inetnum, change sponsoring-org with lir mntner is not possible"() {
         given:
         syncUpdate(getTransient("ASSIGN-PI") + "password: hm\npassword: lir\npassword: owner3")
@@ -377,44 +279,6 @@ class AllocationAttributeValidationSpec extends BaseQueryUpdateSpec {
         ack.errorMessagesFor("Modify", "[inetnum] 192.168.255.0 - 192.168.255.255") == [
                 "The \"sponsoring-org\" attribute can only be changed by the RIPE NCC"]
     }
-
-    def "modify inetnum, change lir-mnt (mnt-by) by lir mntner is not possible"() {
-        given:
-        syncUpdate(getTransient("ALLOC-PA-MANDATORY") + "override: denis, override1")
-
-        expect:
-        queryObject("-GBr -T inetnum 192.168.0.0 - 192.169.255.255", "inetnum", "192.168.0.0 - 192.169.255.255")
-
-        when:
-        def ack = syncUpdateWithResponse("""
-                inetnum:      192.168.0.0 - 192.169.255.255
-                netname:      TEST-NET-NAME
-                country:      NL
-                org:          ORG-LIR1-TEST
-                admin-c:      TP1-TEST
-                tech-c:       TP1-TEST
-                status:       ALLOCATED PA
-                mnt-by:       RIPE-NCC-HM-MNT
-                mnt-by:       LIR2-MNT
-                source:       TEST
-                password: lir
-                """.stripIndent()
-        )
-
-        then:
-
-        ack.errors
-        ack.summary.nrFound == 1
-        ack.summary.assertSuccess(0, 0, 0, 0, 0)
-        ack.summary.assertErrors(1, 0, 1, 0)
-
-        ack.countErrorWarnInfo(1, 0, 0)
-        ack.errors.any { it.operation == "Modify" && it.key == "[inetnum] 192.168.0.0 - 192.169.255.255" }
-
-        ack.errorMessagesFor("Modify", "[inetnum] 192.168.0.0 - 192.169.255.255") == [
-                "Attribute \"mnt-by:\" can only be changed by the RIPE NCC for this object. Please contact \"ncc@ripe.net\" to change it."]
-    }
-
 
     def "modify inetnum, change netname with lir mnt is possible for non-toplevel allocations"() {
         given:
@@ -497,135 +361,6 @@ class AllocationAttributeValidationSpec extends BaseQueryUpdateSpec {
                 "The \"netname\" attribute can only be changed by the RIPE NCC"]
     }
 
-    def "modify inetnum, cannot change (mnt-lower) ripe-ncc-hm-mnt by lir"() {
-        given:
-        syncUpdate(getTransient("ALLOC-PA-RIPE-NCC-MNTNER") + "override: denis, override1")
-        syncUpdate(getTransient("IRT") + "override: denis, override1")
-
-        expect:
-        queryObject("-GBr -T inetnum 192.168.0.0 - 192.169.255.255", "inetnum", "192.168.0.0 - 192.169.255.255")
-        queryObject("-r -T irt irt-test", "irt", "irt-test")
-
-        when:
-        def ack = syncUpdateWithResponse("""
-                inetnum:      192.168.0.0 - 192.169.255.255
-                netname:      TEST-NET-NAME
-                country:      NL
-                org:          ORG-LIR1-TEST
-                admin-c:      TP1-TEST
-                tech-c:       TP1-TEST
-                status:       ALLOCATED PA
-                mnt-by:       RIPE-NCC-HM-MNT
-                mnt-by:       LIR-MNT
-                mnt-lower:    LIR2-MNT          # changed
-                mnt-routes:   RIPE-NCC-HM-MNT
-                mnt-domains:  RIPE-NCC-HM-MNT
-                source:       TEST
-                password: lir
-                password: irt
-                """.stripIndent()
-        )
-
-        then:
-        ack.errors
-
-        ack.summary.nrFound == 1
-        ack.summary.assertSuccess(0, 0, 0, 0, 0)
-        ack.summary.assertErrors(1, 0, 1, 0)
-
-        ack.countErrorWarnInfo(1, 0, 0)
-        ack.errors.any { it.operation == "Modify" && it.key == "[inetnum] 192.168.0.0 - 192.169.255.255" }
-        ack.errorMessagesFor("Modify", "[inetnum] 192.168.0.0 - 192.169.255.255") == [
-                "Adding or removing a RIPE NCC maintainer requires administrative authorisation"
-        ]
-    }
-
-    def "modify inetnum, cannot change (mnt-routes) ripe-ncc-hm-mnt by lir"() {
-        given:
-        syncUpdate(getTransient("ALLOC-PA-RIPE-NCC-MNTNER") + "override: denis, override1")
-        syncUpdate(getTransient("IRT") + "override: denis, override1")
-
-        expect:
-        queryObject("-GBr -T inetnum 192.168.0.0 - 192.169.255.255", "inetnum", "192.168.0.0 - 192.169.255.255")
-        queryObject("-r -T irt irt-test", "irt", "irt-test")
-
-        when:
-        def ack = syncUpdateWithResponse("""
-                inetnum:      192.168.0.0 - 192.169.255.255
-                netname:      TEST-NET-NAME
-                country:      NL
-                org:          ORG-LIR1-TEST
-                admin-c:      TP1-TEST
-                tech-c:       TP1-TEST
-                status:       ALLOCATED PA
-                mnt-by:       RIPE-NCC-HM-MNT
-                mnt-by:       LIR-MNT
-                mnt-lower:    RIPE-NCC-HM-MNT
-                mnt-routes:   LIR2-MNT          # changed
-                mnt-domains:  RIPE-NCC-HM-MNT
-                source:       TEST
-                password: lir
-                password: irt
-                """.stripIndent()
-        )
-
-        then:
-        ack.errors
-
-        ack.summary.nrFound == 1
-        ack.summary.assertSuccess(0, 0, 0, 0, 0)
-        ack.summary.assertErrors(1, 0, 1, 0)
-
-        ack.countErrorWarnInfo(1, 0, 0)
-        ack.errors.any { it.operation == "Modify" && it.key == "[inetnum] 192.168.0.0 - 192.169.255.255" }
-        ack.errorMessagesFor("Modify", "[inetnum] 192.168.0.0 - 192.169.255.255") == [
-                "Adding or removing a RIPE NCC maintainer requires administrative authorisation"
-        ]
-    }
-
-    def "modify inetnum, cannot change (mnt-domains) ripe-ncc-hm-mnt by lir"() {
-        given:
-        syncUpdate(getTransient("ALLOC-PA-RIPE-NCC-MNTNER") + "override: denis, override1")
-        syncUpdate(getTransient("IRT") + "override: denis, override1")
-
-        expect:
-        queryObject("-GBr -T inetnum 192.168.0.0 - 192.169.255.255", "inetnum", "192.168.0.0 - 192.169.255.255")
-        queryObject("-r -T irt irt-test", "irt", "irt-test")
-
-        when:
-        def ack = syncUpdateWithResponse("""
-                inetnum:      192.168.0.0 - 192.169.255.255
-                netname:      TEST-NET-NAME
-                country:      NL
-                org:          ORG-LIR1-TEST
-                admin-c:      TP1-TEST
-                tech-c:       TP1-TEST
-                status:       ALLOCATED PA
-                mnt-by:       RIPE-NCC-HM-MNT
-                mnt-by:       LIR-MNT
-                mnt-lower:    RIPE-NCC-HM-MNT
-                mnt-routes:   RIPE-NCC-HM-MNT
-                mnt-domains:  LIR2-MNT          # changed
-                source:       TEST
-                password: lir
-                password: irt
-                """.stripIndent()
-        )
-
-        then:
-        ack.errors
-
-        ack.summary.nrFound == 1
-        ack.summary.assertSuccess(0, 0, 0, 0, 0)
-        ack.summary.assertErrors(1, 0, 1, 0)
-
-        ack.countErrorWarnInfo(1, 0, 0)
-        ack.errors.any { it.operation == "Modify" && it.key == "[inetnum] 192.168.0.0 - 192.169.255.255" }
-        ack.errorMessagesFor("Modify", "[inetnum] 192.168.0.0 - 192.169.255.255") == [
-                "Adding or removing a RIPE NCC maintainer requires administrative authorisation"
-        ]
-    }
-
     ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
     //
     //  LIR *CHANGE* attributes (inetnum and inet6num)  WITH OVERRIDE
@@ -668,42 +403,6 @@ class AllocationAttributeValidationSpec extends BaseQueryUpdateSpec {
         ack.countErrorWarnInfo(0, 0, 1)
         ack.successes.any { it.operation == "Modify" && it.key == "[inetnum] 192.168.0.0 - 192.169.255.255" }
 
-    }
-
-    def "modify inetnum, change org with override is possible"() {
-        given:
-        syncUpdate(getTransient("ALLOC-PA-MANDATORY") + "password: hm\npassword: owner3")
-
-        expect:
-        queryObject("-GBr -T inetnum 192.168.0.0 - 192.169.255.255", "inetnum", "192.168.0.0 - 192.169.255.255")
-
-        when:
-        def ack = syncUpdateWithResponse("""
-                inetnum:      192.168.0.0 - 192.169.255.255
-                netname:      TEST-NET-NAME
-                descr:        TEST network
-                country:      NL
-                org:          ORG-LIR2-TEST
-                admin-c:      TP1-TEST
-                tech-c:       TP1-TEST
-                status:       ALLOCATED PA
-                mnt-by:       RIPE-NCC-HM-MNT
-                mnt-by:       LIR-MNT
-                mnt-lower:    LIR-MNT
-                source:       TEST
-                override: denis,override1
-                """.stripIndent()
-        )
-
-        then:
-        ack.success
-
-        ack.summary.nrFound == 1
-        ack.summary.assertSuccess(1, 0, 1, 0, 0)
-        ack.summary.assertErrors(0, 0, 0, 0)
-
-        ack.countErrorWarnInfo(0, 0, 1)
-        ack.successes.any { it.operation == "Modify" && it.key == "[inetnum] 192.168.0.0 - 192.169.255.255" }
     }
 
     def "modify inetnum, change sponsoring-org with override is possible"() {
@@ -778,136 +477,6 @@ class AllocationAttributeValidationSpec extends BaseQueryUpdateSpec {
 
         ack.countErrorWarnInfo(0, 0, 1)
         ack.successes.any { it.operation == "Modify" && it.key == "[inetnum] 192.168.0.0 - 192.169.255.255" }
-    }
-
-    def "modify inetnum, change netname with override is possible"() {
-        given:
-        syncUpdate(getTransient("ALLOC-PA-MANDATORY") + "password: hm\npassword: owner3")
-
-        expect:
-        queryObject("-GBr -T inetnum 192.168.0.0 - 192.169.255.255", "inetnum", "192.168.0.0 - 192.169.255.255")
-
-        when:
-        def ack = syncUpdateWithResponse("""
-                inetnum:      192.168.0.0 - 192.169.255.255
-                netname:      DIFFERENT-TEST-NET-NAME
-                descr:        TEST network
-                country:      NL
-                org:          ORG-LIR1-TEST
-                admin-c:      TP1-TEST
-                tech-c:       TP1-TEST
-                status:       ALLOCATED PA
-                mnt-by:       RIPE-NCC-HM-MNT
-                mnt-by:       LIR-MNT
-                mnt-lower:    LIR-MNT
-                source:       TEST
-                override: denis,override1
-                """.stripIndent()
-        )
-
-        then:
-        ack.success
-
-        ack.summary.nrFound == 1
-        ack.summary.assertSuccess(1, 0, 1, 0, 0)
-        ack.summary.assertErrors(0, 0, 0, 0)
-
-        ack.countErrorWarnInfo(0, 0, 1)
-        ack.successes.any { it.operation == "Modify" && it.key == "[inetnum] 192.168.0.0 - 192.169.255.255" }
-    }
-
-    ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-    //
-    //  LIR *CHANGE* attributes (inetnum and inet6num)  WITH RS PASSWORD
-    //
-    ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-
-    def "modify inetnum, updated (all) lir-unlocked attributes"() {
-        given:
-        syncUpdate(getTransient("IRT") + "override: denis, override1")
-        syncUpdate(getTransient("IRT2") + "override: denis, override1")
-        syncUpdate(getTransient("DOMAINS2-MNT") + "override: denis, override1")
-        syncUpdate(getTransient("ALLOC-PA-EXTRA") + "override: denis, override1")
-
-        expect:
-        queryObject("-GBr -T inetnum 192.168.0.0 - 192.169.255.255", "inetnum", "192.168.0.0 - 192.169.255.255")
-        queryObject("-r -T mntner DOMAINS2-MNT", "mntner", "DOMAINS2-MNT")
-        queryObject("-r -T irt irt-test", "irt", "irt-test")
-        queryObject("-r -T irt IRT-2-TEST", "irt", "IRT-2-TEST")
-
-        when:
-        def ack = syncUpdateWithResponse("""
-                inetnum:      192.168.0.0 - 192.169.255.255
-                netname:      TEST-NET-NAME
-                descr:        other description # changed
-                country:      DE                # changed
-                geoloc:       9.0 9.0           # changed
-                language:     DE                # changed
-                org:          ORG-LIR1-TEST
-                admin-c:      TP2-TEST          # changed
-                tech-c:       TP2-TEST          # changed
-                status:       ALLOCATED PA
-                mnt-by:       RIPE-NCC-HM-MNT
-                mnt-by:       LIR-MNT
-                remarks:      a different remark# changed
-                notify:       other@ripe.net    # changed
-                mnt-lower:    LIR2-MNT          # changed
-                mnt-routes:   OWNER2-MNT        # changed
-                mnt-domains:  DOMAINS-MNT       # changed
-                mnt-irt:      IRT-2-TEST        # changed
-                source:       TEST
-                password: lir
-                password: irt
-                """.stripIndent()
-        )
-
-        then:
-        ack.success
-
-        ack.summary.nrFound == 1
-        ack.summary.assertSuccess(1, 0, 1, 0, 0)
-        ack.summary.assertErrors(0, 0, 0, 0)
-        ack.countErrorWarnInfo(0, 0, 0)
-        ack.successes.any { it.operation == "Modify" && it.key == "[inetnum] 192.168.0.0 - 192.169.255.255" }
-    }
-
-
-    def "modify inetnum, change mnt-by with rs password should be possible"() {
-        given:
-        syncUpdate(getTransient("ALLOC-PA-MANDATORY") + "password: hm\npassword: owner3")
-
-        expect:
-        queryObject("-GBr -T inetnum 192.168.0.0 - 192.169.255.255", "inetnum", "192.168.0.0 - 192.169.255.255")
-
-        when:
-        def ack = syncUpdateWithResponse("""
-                inetnum:      192.168.0.0 - 192.169.255.255
-                netname:      TEST-NET-NAME
-                descr:        TEST network
-                country:      NL
-                org:          ORG-LIR1-TEST
-                admin-c:      TP1-TEST
-                tech-c:       TP1-TEST
-                status:       ALLOCATED PA
-                mnt-by:       RIPE-NCC-HM-MNT
-                mnt-by:       LIR2-MNT
-                mnt-lower:    LIR-MNT
-                source:       TEST
-                password:     hm
-                """.stripIndent()
-        )
-
-        then:
-        ack.success
-
-        ack.summary.nrFound == 1
-        ack.summary.assertSuccess(1, 0, 1, 0, 0)
-        ack.summary.assertErrors(0, 0, 0, 0)
-
-        ack.countErrorWarnInfo(0, 0, 0)
-        ack.successes.any { it.operation == "Modify" && it.key == "[inetnum] 192.168.0.0 - 192.169.255.255" }
-
     }
 
     def "modify inetnum, change org with RS mntner is possible"() {
@@ -1027,299 +596,5 @@ class AllocationAttributeValidationSpec extends BaseQueryUpdateSpec {
 
     ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-    def "modify inetnum, change status with RS mntner and lir mnt authed is not possible"() {
-        given:
-        syncUpdate(getTransient("ALLOC-PA-MANDATORY") + "password: hm\npassword: owner3")
 
-        expect:
-        queryObject("-GBr -T inetnum 192.168.0.0 - 192.169.255.255", "inetnum", "192.168.0.0 - 192.169.255.255")
-
-        when:
-        def ack = syncUpdateWithResponse("""
-                inetnum:      192.168.0.0 - 192.169.255.255
-                netname:      TEST-NET-NAME
-                descr:        TEST network
-                country:      NL
-                org:          ORG-LIR1-TEST
-                admin-c:      TP1-TEST
-                tech-c:       TP1-TEST
-                status:       ASSIGNED PA
-                mnt-by:       RIPE-NCC-HM-MNT
-                mnt-by:       LIR-MNT
-                mnt-lower:    LIR-MNT
-                source:       TEST
-
-                password: hm
-                password: lir
-                """.stripIndent()
-        )
-
-        then:
-
-        ack.errors
-        ack.summary.nrFound == 1
-        ack.summary.assertSuccess(0, 0, 0, 0, 0)
-        ack.summary.assertErrors(1, 0, 1, 0)
-
-        ack.countErrorWarnInfo(1, 0, 0)
-        ack.errors.any { it.operation == "Modify" && it.key == "[inetnum] 192.168.0.0 - 192.169.255.255" }
-
-        ack.errorMessagesFor("Modify", "[inetnum] 192.168.0.0 - 192.169.255.255") == [
-                "status value cannot be changed, you must delete and re-create the object"]
-    }
-
-    def "modify inetnum, delete (all) lir-unlocked attributes by lir"() {
-        given:
-        syncUpdate(getTransient("IRT") + "override: denis, override1")
-        syncUpdate(getTransient("ALLOC-PA-EXTRA") + "override: denis, override1")
-
-        expect:
-        queryObject("-GBr -T inetnum 192.168.0.0 - 192.169.255.255", "inetnum", "192.168.0.0 - 192.169.255.255")
-        queryObject("-r -T irt irt-test", "irt", "irt-test")
-
-        when:
-        //        descr:        other description # deleted
-        //        geoloc:       9.0 9.0           # deleted
-        //        language:     DE                # deleted
-        //        admin-c:      TP2-TEST          # deleted
-        //        tech-c:       TP2-TEST          # deleted
-        //        remarks:      a different remark# deleted
-        //        notify:       other@ripe.net    # deleted
-        //        mnt-lower:    LIR2-MNT          # deleted
-        //        mnt-routes:   OWNER2-MNT        # deleted
-        //        mnt-domains:  DOMAINS-MNT       # deleted
-        //        mnt-irt:      IRT-2-TEST        # deleted
-        def ack = syncUpdateWithResponse("""
-                inetnum:      192.168.0.0 - 192.169.255.255
-                netname:      TEST-NET-NAME
-                country:      NL
-                org:          ORG-LIR1-TEST
-                admin-c:      TP1-TEST
-                tech-c:       TP1-TEST
-                status:       ALLOCATED PA
-                mnt-by:       RIPE-NCC-HM-MNT
-                mnt-by:       LIR-MNT
-                source:       TEST
-                password: lir
-                """.stripIndent()
-        )
-
-        then:
-        ack.success
-
-        ack.summary.nrFound == 1
-        ack.summary.assertSuccess(1, 0, 1, 0, 0)
-        ack.summary.assertErrors(0, 0, 0, 0)
-        ack.countErrorWarnInfo(0, 0, 0)
-        ack.successes.any { it.operation == "Modify" && it.key == "[inetnum] 192.168.0.0 - 192.169.255.255" }
-    }
-
-    def "modify inetnum, cannot delete (some) lir-unlocked attributes by lir"() {
-        given:
-        syncUpdate(getTransient("ALLOC-PA-MANDATORY") + "override: denis, override1")
-
-        expect:
-        queryObject("-GBr -T inetnum 192.168.0.0 - 192.169.255.255", "inetnum", "192.168.0.0 - 192.169.255.255")
-
-        when:
-        //        org:          ORG-LIR1-TEST # cannot delete, but warning is NOT presented!!
-        //        country:      NL            # cannot delete
-        //        admin-c:      TP1-TEST      # cannot delete
-        //        tech-c:       TP1-TEST      # cannot delete
-        def ack = syncUpdateWithResponse("""
-                inetnum:      192.168.0.0 - 192.169.255.255
-                netname:      TEST-NET-NAME
-                status:       ALLOCATED PA
-                mnt-by:       RIPE-NCC-HM-MNT
-                mnt-by:       LIR-MNT
-                source:       TEST
-                password: lir
-                """.stripIndent()
-        )
-
-        then:
-        ack.errors
-
-        ack.summary.nrFound == 1
-        ack.summary.assertSuccess(0, 0, 0, 0, 0)
-        ack.summary.assertErrors(1, 0, 1, 0)
-        ack.countErrorWarnInfo(4, 0, 0)
-
-        ack.errorMessagesFor("Modify", "[inetnum] 192.168.0.0 - 192.169.255.255") == [
-                "Mandatory attribute \"country\" is missing",
-                "Mandatory attribute \"admin-c\" is missing",
-                "Mandatory attribute \"tech-c\" is missing"]
-    }
-
-    def "modify inetnum, cannot delete (org) lir-unlocked attributes by lir"() {
-        given:
-        syncUpdate(getTransient("ALLOC-PA-MANDATORY") + "override: denis, override1")
-
-        expect:
-        queryObject("-GBr -T inetnum 192.168.0.0 - 192.169.255.255", "inetnum", "192.168.0.0 - 192.169.255.255")
-
-        when:
-        //        org:          ORG-LIR1-TEST # cannot delete
-        def ack = syncUpdateWithResponse("""
-                inetnum:      192.168.0.0 - 192.169.255.255
-                netname:      TEST-NET-NAME
-                country:      NL
-                admin-c:      TP1-TEST
-                tech-c:       TP1-TEST
-                status:       ALLOCATED PA
-                mnt-by:       RIPE-NCC-HM-MNT
-                mnt-by:       LIR-MNT
-                source:       TEST
-                password: lir
-                """.stripIndent()
-        )
-
-        then:
-        ack.errors
-
-        ack.summary.nrFound == 1
-        ack.summary.assertSuccess(0, 0, 0, 0, 0)
-        ack.summary.assertErrors(1, 0, 1, 0)
-
-        ack.countErrorWarnInfo(2, 0, 0)
-        ack.errorMessagesFor("Modify", "[inetnum] 192.168.0.0 - 192.169.255.255") == [
-                "Referenced organisation can only be removed by the RIPE NCC for this resource. Please contact \"ncc@ripe.net\" to remove this reference.",
-                "Missing required \"org:\" attribute"]
-    }
-
-    def "modify inetnum, cannot delete (mnt-lower) ripe-ncc-hm-mnt by lir"() {
-        given:
-        syncUpdate(getTransient("ALLOC-PA-EXTRA-RIPE-NCC-MNTNER") + "override: denis, override1")
-        syncUpdate(getTransient("IRT") + "override: denis, override1")
-
-        expect:
-        queryObject("-GBr -T inetnum 192.168.0.0 - 192.169.255.255", "inetnum", "192.168.0.0 - 192.169.255.255")
-        queryObject("-r -T irt irt-test", "irt", "irt-test")
-
-        when:
-        def ack = syncUpdateWithResponse("""
-                inetnum:      192.168.0.0 - 192.169.255.255
-                netname:      TEST-NET-NAME
-                country:      NL
-                org:          ORG-LIR1-TEST
-                admin-c:      TP1-TEST
-                tech-c:       TP1-TEST
-                status:       ALLOCATED PA
-                mnt-by:       RIPE-NCC-HM-MNT
-                mnt-by:       LIR-MNT
-                mnt-lower:    RIPE-NCC-HM-MNT  # hm-mnt
-                mnt-lower:    LIR-MNT          # extra
-                mnt-routes:   RIPE-NCC-HM-MNT  # hm-mnt
-                mnt-routes:   OWNER-MNT        # extra
-                mnt-domains:  RIPE-NCC-HM-MNT  # hm-mnt
-                mnt-domains:  DOMAINS-MNT      # extra
-                source:       TEST
-                password: lir
-                password: irt
-                """.stripIndent()
-        )
-
-        then:
-        ack.errors
-
-        ack.summary.nrFound == 1
-        ack.summary.assertSuccess(0, 0, 0, 0, 0)
-        ack.summary.assertErrors(1, 0, 1, 0)
-
-        ack.countErrorWarnInfo(1, 0, 0)
-        ack.errors.any { it.operation == "Modify" && it.key == "[inetnum] 192.168.0.0 - 192.169.255.255" }
-        ack.errorMessagesFor("Modify", "[inetnum] 192.168.0.0 - 192.169.255.255") == [
-                "Adding or removing a RIPE NCC maintainer requires administrative authorisation"
-        ]
-    }
-
-    def "modify inetnum, cannot delete (mnt-routes) ripe-ncc-hm-mnt by lir"() {
-        given:
-        syncUpdate(getTransient("ALLOC-PA-EXTRA-RIPE-NCC-MNTNER") + "override: denis, override1")
-        syncUpdate(getTransient("IRT") + "override: denis, override1")
-
-        expect:
-        queryObject("-GBr -T inetnum 192.168.0.0 - 192.169.255.255", "inetnum", "192.168.0.0 - 192.169.255.255")
-        queryObject("-r -T irt irt-test", "irt", "irt-test")
-
-        when:
-        def ack = syncUpdateWithResponse("""
-                inetnum:      192.168.0.0 - 192.169.255.255
-                netname:      TEST-NET-NAME
-                country:      NL
-                org:          ORG-LIR1-TEST
-                admin-c:      TP1-TEST
-                tech-c:       TP1-TEST
-                status:       ALLOCATED PA
-                mnt-by:       RIPE-NCC-HM-MNT
-                mnt-by:       LIR-MNT
-                mnt-lower:    RIPE-NCC-HM-MNT  # hm-mnt
-                mnt-lower:    LIR-MNT          # extra
-                mnt-routes:   OWNER-MNT        # extra
-                mnt-domains:  RIPE-NCC-HM-MNT  # hm-mnt
-                mnt-domains:  DOMAINS-MNT      # extra
-                source:       TEST
-                password: lir
-                password: irt
-                """.stripIndent()
-        )
-
-        then:
-        ack.errors
-
-        ack.summary.nrFound == 1
-        ack.summary.assertSuccess(0, 0, 0, 0, 0)
-        ack.summary.assertErrors(1, 0, 1, 0)
-
-        ack.countErrorWarnInfo(1, 0, 0)
-        ack.errors.any { it.operation == "Modify" && it.key == "[inetnum] 192.168.0.0 - 192.169.255.255" }
-        ack.errorMessagesFor("Modify", "[inetnum] 192.168.0.0 - 192.169.255.255") == [
-                "Adding or removing a RIPE NCC maintainer requires administrative authorisation"
-        ]
-    }
-
-    def "modify inetnum, cannot delete (mnt-domains) ripe-ncc-hm-mnt by lir"() {
-        given:
-        syncUpdate(getTransient("ALLOC-PA-EXTRA-RIPE-NCC-MNTNER") + "override: denis, override1")
-        syncUpdate(getTransient("IRT") + "override: denis, override1")
-
-        expect:
-        queryObject("-GBr -T inetnum 192.168.0.0 - 192.169.255.255", "inetnum", "192.168.0.0 - 192.169.255.255")
-        queryObject("-r -T irt irt-test", "irt", "irt-test")
-
-        when:
-        def ack = syncUpdateWithResponse("""
-                inetnum:      192.168.0.0 - 192.169.255.255
-                netname:      TEST-NET-NAME
-                country:      NL
-                org:          ORG-LIR1-TEST
-                admin-c:      TP1-TEST
-                tech-c:       TP1-TEST
-                status:       ALLOCATED PA
-                mnt-by:       RIPE-NCC-HM-MNT
-                mnt-by:       LIR-MNT
-                mnt-lower:    RIPE-NCC-HM-MNT  # hm-mnt
-                mnt-lower:    LIR-MNT          # extra
-                mnt-routes:   RIPE-NCC-HM-MNT  # hm-mnt
-                mnt-routes:   OWNER-MNT        # extra
-                mnt-domains:  DOMAINS-MNT
-                source:       TEST
-                password: lir
-                password: irt
-                """.stripIndent()
-        )
-
-        then:
-        ack.errors
-
-        ack.summary.nrFound == 1
-        ack.summary.assertSuccess(0, 0, 0, 0, 0)
-        ack.summary.assertErrors(1, 0, 1, 0)
-
-        ack.countErrorWarnInfo(1, 0, 0)
-        ack.errors.any { it.operation == "Modify" && it.key == "[inetnum] 192.168.0.0 - 192.169.255.255" }
-        ack.errorMessagesFor("Modify", "[inetnum] 192.168.0.0 - 192.169.255.255") == [
-                "Adding or removing a RIPE NCC maintainer requires administrative authorisation"
-        ]
-    }
 }
